Add test case with control character
The mox-worker item #3656 happens because there is a \x13 control character in the source map json.
It happens when the angular-1.4.5.js source content is included in the source map, but strangely in the
source map there is a \x13 character on the position where there is a \xe2\x80\x93 in the original source.

In the test case I added a js source containing both characters, minified it by
`uglifyjs control_character.js -m --source-map includeSources -o control_character.min.js`
and load the source map as a json. It succeeds, because the minification escaped the control character properly.